add an 'age_check' parameter to root_cache
Add this so that I can turn it off root cache aging out while running
the mock tests. Speeds things up on a slow link.

+++ py/mockbuild/plugins/root_cache.py
@@ -75,19 +75,22 @@ class RootCache(object):
 
         # check cache status
         try:
-            # see if it aged out
-            statinfo = os.stat(self.rootCacheFile)
-            file_age_days = (time.time() - statinfo.st_ctime) / (60 * 60 * 24)
-            if file_age_days > self.root_cache_opts['max_age_days']:
-                getLog().info("root cache aged out! cache will be rebuilt")
-                os.unlink(self.rootCacheFile)
+            if self.root_cache_opts['age_check']:
+                # see if it aged out
+                statinfo = os.stat(self.rootCacheFile)
+                file_age_days = (time.time() - statinfo.st_ctime) / (60 * 60 * 24)
+                if file_age_days > self.root_cache_opts['max_age_days']:
+                    getLog().info("root cache aged out! cache will be rebuilt")
+                    os.unlink(self.rootCacheFile)
+                else:
+                    # make sure no config file is newer than the cache file
+                    for cfg in self.rootObj.configs:
+                        if os.stat(cfg).st_mtime > statinfo.st_mtime:
+                            getLog().info("%s newer than root cache; cache will be rebuilt" % cfg)
+                            os.unlink(self.rootCacheFile)
+                            break
             else:
-                # make sure no config file is newer than the cache file
-                for cfg in self.rootObj.configs:
-                    if os.stat(cfg).st_mtime > statinfo.st_mtime:
-                        getLog().info("%s newer than root cache; cache will be rebuilt" % cfg)
-                        os.unlink(self.rootCacheFile)
-                        break
+                getLog().info("skipping root_cache aging check")
         except OSError:
             pass
